Pagina_web/BACKEND/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
crytography
pymysql
python-dotenv